add struct intset and interface set

+++ datastruct/set/intset_test.go
@@ -7,7 +7,7 @@ import (
 
 func TestAdd(t *testing.T) {
 	intSet := NewIntSet()
-	intSet.Add(100)
-	intSet.Add(math.MaxInt16 + 1)
-	intSet.Add(math.MaxInt32 + 1)
+	intSet.Add(int64(100))
+	intSet.Add(int64(math.MaxInt16 + 1))
+	intSet.Add(int64(math.MaxInt32 + 1))
 }
